Leave `xt/` tests where they are, don't move them to `t/`.
Now that I've started building more Dists w/this Author Bundle, I'm
finding that the test output is easier to read when the `xt/` tests are
left where they are.

Also helps make it clearer in the final release tarball that those are
"extra tests" and aren't part of the regular test suite.