Fix handling of `script/` dirs, for Static Install.  Fixes #1
Although we were using the `StaticInstall` DZil plugin to determine that
we are eligible for doing a Static Install, we were not indicating that
scripts to install should be found in the `script/` directory.

So... check for presence of that directory, and mark that as the
`ExecDir` properly.  Otherwise, fall back to `bin/` (the default).